fix: rename github icon
